requirements.txt
fastapi==0.99.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.18.0
streamlit==1.29.0
faker==19.6.2
prometheus-client==0.19.0
networkx==3.2.1
xgboost==2.0.3
prophet==1.1.5
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
faiss-cpu==1.7.4
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.34.0
ollama==0.1.6

# LangChain and related
langchain==0.0.346
openai==0.28.1
tiktoken==0.5.1
chromadb==0.4.10

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.3
pydantic>=1.10.0
pyyaml>=6.0


# Utilities
python-dotenv==1.0.0